Makes the code consistent for both http and rpc
